pyproject.toml
[tool.poetry]
name = "marinaalchemist"
version = "0.1.0"
description = ""
authors = ["Ajay"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.2"
allure-pytest = "^2.13.2"
pandas = "^2.0.3"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
